+++ toolkit/offchain/src/csl.rs
@@ -321,7 +321,7 @@ pub(crate) trait TransactionBuilderExt {
 	fn add_mint_one_script_token(
 		&mut self,
 		script: &PlutusScript,
-		ex_units: ExUnits,
+		ex_units: &ExUnits,
 	) -> Result<(), JsError>;
 
 	/// Adds minting of 1 token (with empty asset name) for the given script using reference input
@@ -329,7 +329,7 @@ pub(crate) trait TransactionBuilderExt {
 		&mut self,
 		script: &PlutusScript,
 		ref_input: &TransactionInput,
-		ex_units: ExUnits,
+		ex_units: &ExUnits,
 	) -> Result<(), JsError>;
 
 	/// Sets fields required by the most of partner-chains smart contract transactions.
@@ -389,7 +389,7 @@ impl TransactionBuilderExt for TransactionBuilder {
 	fn add_mint_one_script_token(
 		&mut self,
 		script: &PlutusScript,
-		ex_units: ExUnits,
+		ex_units: &ExUnits,
 	) -> Result<(), JsError> {
 		let mut mint_builder = self.get_mint_builder().unwrap_or(MintBuilder::new());
 
@@ -400,7 +400,7 @@ impl TransactionBuilderExt for TransactionBuilder {
 				&RedeemerTag::new_mint(),
 				&0u32.into(),
 				&PlutusData::new_empty_constr_plutus_data(&0u32.into()),
-				&ex_units,
+				ex_units,
 			),
 		);
 		mint_builder.add_asset(&mint_witness, &empty_asset_name(), &Int::new_i32(1))?;
@@ -412,7 +412,7 @@ impl TransactionBuilderExt for TransactionBuilder {
 		&mut self,
 		script: &PlutusScript,
 		ref_input: &TransactionInput,
-		ex_units: ExUnits,
+		ex_units: &ExUnits,
 	) -> Result<(), JsError> {
 		let mut mint_builder = self.get_mint_builder().unwrap_or(MintBuilder::new());
 
@@ -428,7 +428,7 @@ impl TransactionBuilderExt for TransactionBuilder {
 				&RedeemerTag::new_mint(),
 				&0u32.into(),
 				&PlutusData::new_empty_constr_plutus_data(&0u32.into()),
-				&ex_units,
+				ex_units,
 			),
 		);
 		mint_builder.add_asset(&mint_witness, &empty_asset_name(), &Int::new_i32(1))?;
@@ -500,7 +500,7 @@ pub(crate) trait InputsBuilderExt: Sized {
 		&mut self,
 		utxo: &OgmiosUtxo,
 		script: &PlutusScript,
-		ex_units: ExUnits,
+		ex_units: &ExUnits,
 	) -> Result<(), JsError>;
 
 	/// Adds ogmios inputs to the tx inputs builder.
@@ -515,7 +515,7 @@ impl InputsBuilderExt for TxInputsBuilder {
 		&mut self,
 		utxo: &OgmiosUtxo,
 		script: &PlutusScript,
-		ex_units: ExUnits,
+		ex_units: &ExUnits,
 	) -> Result<(), JsError> {
 		let input = utxo.to_csl_tx_input();
 		let amount = convert_value(&utxo.value)?;
@@ -526,7 +526,7 @@ impl InputsBuilderExt for TxInputsBuilder {
 				// CSL will set redeemer index for the index of script input after sorting transaction inputs
 				&0u32.into(),
 				&PlutusData::new_empty_constr_plutus_data(&0u32.into()),
-				&ex_units,
+				ex_units,
 			),
 		);
 		self.add_plutus_script_input(&witness, &input, &amount);
@@ -729,11 +729,12 @@ mod tests {
 
 #[cfg(test)]
 mod prop_tests {
-	use super::{get_builder_config, OgmiosUtxoExt, TransactionBuilderExt, TransactionContext};
+	use super::{
+		get_builder_config, zero_ex_units, OgmiosUtxoExt, TransactionBuilderExt, TransactionContext,
+	};
 	use crate::test_values::*;
 	use cardano_serialization_lib::{
-		BigNum, ExUnits, NetworkIdKind, Transaction, TransactionBuilder, TransactionInputs,
-		TransactionOutput, Value,
+		NetworkIdKind, Transaction, TransactionBuilder, TransactionInputs, TransactionOutput, Value,
 	};
 	use ogmios_client::types::OgmiosValue;
 	use ogmios_client::types::{OgmiosTx, OgmiosUtxo};
@@ -755,12 +756,7 @@ mod prop_tests {
 			protocol_parameters: protocol_parameters(),
 		};
 		let mut tx_builder = TransactionBuilder::new(&get_builder_config(&ctx).unwrap());
-		tx_builder
-			.add_mint_one_script_token(
-				&test_policy(),
-				ExUnits::new(&BigNum::zero(), &BigNum::zero()),
-			)
-			.unwrap();
+		tx_builder.add_mint_one_script_token(&test_policy(), &zero_ex_units()).unwrap();
 		tx_builder
 			.add_output_with_one_script_token(
 				&test_validator(),