+++ changelog.md
@@ -6,7 +6,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 ## Changed
 
-* `setup-main-chain-state` command now uses native Rust to insert the D-Parameter
+* `setup-main-chain-state` command now uses native Rust to upsert the D-Parameter and upsert permissioned candidates
 
 ## Removed